Logger folder was renamed to Loggers
